move omarchy shell commands to its own file
