+++ src-tauri/src/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod hyprland;
 pub mod settings;
 pub mod shell;
 pub mod update_config;
+pub mod omarchy;
